refactoring to make it a bit more organized
