+++ src/components/content/conditions/patient-conditions-actions.tsx
@@ -4,8 +4,8 @@ import { useState } from "react";
 import {
   conditionAddSchema,
   getAddConditionData,
-} from "../forms/condition-schema";
-import { createOrEditCondition } from "../forms/conditions";
+} from "../forms/schemas/condition-schema";
+import { createOrEditCondition } from "../forms/actions/conditions";
 import { DrawerFormWithFields } from "@/components/core/form/drawer-form-with-fields";
 import { usePatient } from "@/components/core/patient-provider";
 import { Toggle } from "@/components/core/toggle";
