+++ src/components/content/forms/actions/medications.tsx
@@ -1,6 +1,3 @@
-import type { FormEntry } from "../../core/form/drawer-form-with-fields";
-import { z } from "zod";
-import { MedicationsAutoComplete } from "./medications-autocomplete";
 import { CTWRequestContext } from "@/components/core/ctw-context";
 import { createOrEditFhirResource } from "@/fhir/action-helper";
 import { dateToISO } from "@/fhir/formatters";
@@ -12,33 +9,6 @@ import {
   QUERY_KEY_PATIENT_MEDICATIONS,
 } from "@/utils/query-keys";
 import { queryClient } from "@/utils/request";
-
-export const medicationStatementSchema = z.object({
-  subjectID: z.string({ required_error: "Patient must be specified." }),
-  dateAsserted: z.date({ required_error: "Date asserted is required." }),
-  medication: z.object({
-    code: z.string({
-      required_error: "Please choose a medication.",
-    }),
-    // These are technically required but we mark them
-    // as optional to avoid duplicative error messages.
-    // The condition autocomplete will set us up so that
-    // all three of these values are set.
-    display: z.string().optional(),
-    system: z.string().optional(),
-  }),
-  dosage: z.string().optional(),
-  status: z.enum([
-    "active",
-    "completed",
-    "entered-in-error",
-    "intended",
-    "not-taken",
-    "on-hold",
-    "stopped",
-    "unknown",
-  ]),
-});
 
 const QUERY_KEYS = [
   QUERY_KEY_PATIENT,
@@ -96,44 +66,3 @@ export const createMedicationStatement = async (
 
   return response;
 };
-
-export const getMedicationFormData = (
-  medication: MedicationStatementModel
-): FormEntry[] => [
-  {
-    label: "Subject",
-    value: medication.subjectID,
-    field: "subjectID",
-    readonly: true,
-    hidden: true,
-  },
-  {
-    label: "Date Asserted",
-    value: medication.dateAsserted,
-    field: "dateAsserted",
-    readonly: true,
-  },
-  {
-    label: "Medication",
-    field: "medication",
-    value: medication.display,
-    readonly: false,
-    render: (readonly: boolean | undefined, inputProps) => (
-      <MedicationsAutoComplete
-        readonly={readonly}
-        {...inputProps}
-        defaultCoding={{}}
-      />
-    ),
-  },
-  {
-    label: "Latest Status",
-    value: medication.status,
-    field: "status",
-  },
-  {
-    label: "Instructions",
-    value: medication.dosage,
-    field: "dosage",
-  },
-];
